pyproject.toml
[project]
name = "beeldensearch"
version = "0.1.0"
description = "CLARIAH Single-Search inspired demo over Openbeelden with multimodal + LLM extensions"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Teng Long", email = "uestc.longteng@gmail.com" }]

dependencies = [
    # Data pipeline
    "httpx>=0.27",
    "lxml>=5.2",
    "tenacity>=9.0",
    "tqdm>=4.66",
    "pillow>=10.3",
    # Backend
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "opensearch-py>=2.6",
    # NER
    "spacy>=3.7",
    # Embeddings / vector
    "open-clip-torch>=2.24",
    "torch>=2.2",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    # LLM
    "litellm>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-tqdm",
]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["data_pipeline*", "backend*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
ignore_missing_imports = true
packages = ["backend", "data_pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests", "data_pipeline/tests"]
